+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.7.0](https://github.com/podlove/podlove-ui/compare/v5.6.4...v5.7.0) (2021-07-11)
+
+
+### Bug Fixes
+
+* **components:** fix issue with mail share in firefox ([875277a](https://github.com/podlove/podlove-ui/commit/875277ac1ff4d6e461eaad78aa6edfb69304cfe7)), closes [#950](https://github.com/podlove/podlove-ui/issues/950)
+
+
+
+
+
+## [5.6.4](https://github.com/podlove/podlove-ui/compare/v5.6.3...v5.6.4) (2021-07-07)
+
+**Note:** Version bump only for package @podlove/utils
+
+
+
+
+
+## [5.6.3](https://github.com/podlove/podlove-ui/compare/v5.6.2...v5.6.3) (2021-07-07)
+
+**Note:** Version bump only for package @podlove/utils
+
+
+
+
+
+## [5.6.2](https://github.com/podlove/podlove-ui/compare/v5.6.1...v5.6.2) (2021-07-03)
+
+**Note:** Version bump only for package @podlove/utils
+
+
+
+
+
+## [5.6.1](https://github.com/podlove/podlove-ui/compare/v5.6.0...v5.6.1) (2021-07-03)
+
+**Note:** Version bump only for package @podlove/utils
+
+
+
+
+
+# [5.6.0](https://github.com/podlove/podlove-ui/compare/v5.5.0...v5.6.0) (2021-06-19)
+
+**Note:** Version bump only for package @podlove/utils
+
+
+
+
+
 # [5.5.0](https://github.com/podlove/podlove-ui/compare/v5.4.1...v5.5.0) (2021-04-05)
 
 **Note:** Version bump only for package @podlove/utils